refactor providers folder
